skip attrs when retrieval fails (e.g., due to decoding failure)
Previously, if ldap_get_values failed, the code immediately following
would still execute, resulting in a segfault.

Note: removed the perror call since it persists on the screen which
makes continued use of the program difficult. It might be better to open
an error log and also to record the failed attr name, however.